Fix docs-test and conditionally run when flag is passed
This avoids running the docs tests e.g. when pulling the package as a
dependency. To run with nix, the derivation needs to be configured with
the flag, otherwise it won't be able to put the dependencies in the
package db. See the `test` target in the Makefile for more info.

This idea was taken from termonad after @cdepillabout kindly answered
my question (on fpslack) about nix+cabal2nix not recognising
build-tool-depends.

[termonad]: https://github.com/cdepillabout/termonad