Add support for GHC 8.8 (#33)
* Add .envrc

* Add missing packages in overrides

* Update nix sources for matrix building

* Fix stack extra-deps for 8.6.5 and allow newer

* Update deprecated `barbies` aliases

* Expose classes from Construct.hs

* Cut new version v0.5.0.0 and update CHANGELOG

* Format docs with ormolu

* Fix docs-test and conditionally run when flag is passed

This avoids running the docs tests e.g. when pulling the package as a
dependency. To run with nix, the derivation needs to be configured with
the flag, otherwise it won't be able to put the dependencies in the
package db. See the `test` target in the Makefile for more info.

This idea was taken from termonad after @cdepillabout kindly answered
my question (on fpslack) about nix+cabal2nix not recognising
build-tool-depends.

[termonad]: https://github.com/cdepillabout/termonad

* Point default stack yaml to 8.8.4